javitas apr5

+++ catalog/catalog.js
@@ -156,9 +156,7 @@ async function loadDocuments(categoryId = null, page = 1) {
 
           loaderContainer.style.display = "none";
 
-          throw new Error(
-            "Kvízgenerálás időtúllépés. Kérlek próbáld újra kevesebb kérdéssel."
-          );
+          throw new Error("Kvízgenerálás időtúllépés. Kérlek próbáld újra.");
         };
 
         const startQuizGeneration = async (lang, maxQuestions) => {
@@ -184,7 +182,7 @@ async function loadDocuments(categoryId = null, page = 1) {
             clearTimeout(timeoutId);
 
             if (!response.ok) {
-              let errorMessage = "Ismeretlen hiba történt.";
+              let errorMessage = "Nem sikerült a generálás.";
               try {
                 const errorData = await response.json();
                 if (errorData.message) {
@@ -374,6 +372,14 @@ async function loadDocuments(categoryId = null, page = 1) {
                 return;
               }
 
+              const confirmed = confirm(
+                "Biztosan le szeretnéd cserélni a fájlt? A régi fájl és a hozzátartozó kvízek, eredmények törlődni fognak."
+              );
+              if (!confirmed) {
+                submitButton.disabled = false;
+                return;
+              }
+
               const formData = new FormData();
               formData.append("uploaded_by", userId);
               formData.append("file", fileNew);
@@ -393,7 +399,8 @@ async function loadDocuments(categoryId = null, page = 1) {
               });
 
               if (!response.ok) {
-                let errorMessage = "100 feltöltés engedélyezett naponta!";
+                let errorMessage =
+                  "Maximum 20 fájlod lehet jóváhagyásra váró státuszban.";
                 try {
                   const errorData = await response.json();
                   errorMessage = errorData.detail || errorMessage;
@@ -473,6 +480,12 @@ async function loadDocuments(categoryId = null, page = 1) {
           };
 
           deleteButton.onclick = async () => {
+            const confirmed = confirm(
+              "Biztosan törölni szeretnéd a fájlt? A fájl és a hozzátartozó kvízek, eredmények törlődni fognak."
+            );
+            if (!confirmed) {
+              return;
+            }
             try {
               const response = await fetch(doc.delete_url, {
                 method: "DELETE",
